move all bot class to bot folerrderr
