+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoTest.java
@@ -2,6 +2,7 @@ package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import org.firstinspires.ftc.teamcode.bot.GyroBot;
 
 
 @Autonomous(name="Auto Test", group="Auto")
