+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CameraTest.java
@@ -2,7 +2,7 @@ package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
+import org.firstinspires.ftc.teamcode.bot.NewCameraBot;
 
 @Autonomous(name="Camera Test", group="Tests")
