+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ManualDriveOpMode.java
@@ -1,9 +1,9 @@
 
-        package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode;
 
-        import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-        import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import org.firstinspires.ftc.teamcode.bot.OdometryBot;
 /**
  * Mecanum teleop (with an optional arcade mode)
  * * Left stick controls x/y translation.
