+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/bot/NewCameraBot.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.bot;
 
 import android.graphics.Bitmap;
 import android.graphics.Color;
